Refactor code for Rental functionnality
