+++ Data.rb
@@ -4,12 +4,18 @@ module Data
 
     @persons = []
 
+    @rentals = []
+
     def self.books
         @books
     end
 
     def self.persons
         @persons
+    end
+
+    def self.rentals
+        @rentals
     end
 
 end
