+++ app.rb
@@ -6,15 +6,16 @@ require './classroom'
 require './rental'
 require './bookLogic'
 require './personLogic'
+require './rentals'
 
 class App
 
-  attr_accessor :books_class, :persons_class
+  attr_accessor :books_class, :persons_class, :rentals
 
   def initialize
     @books_class = BookLogic.new
     @persons_class = PersonLogic.new
-    @rentals = []
+    @rentals = Rentals.new
   end
 
   def start_console
@@ -27,44 +28,6 @@ class App
       end
 
       option input
-    end
-  end
-
-  def create_rental
-    puts 'Select which book you want to rent by entering its number'
-    @books.each_with_index { |book, index| puts "#{index}) Title: #{book.title}, Author: #{book.author}" }
-
-    book_id = gets.chomp.to_i
-
-    puts 'Select a person from the list by its number'
-    @persons.each_with_index do |person, index|
-      puts "#{index}) [#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-
-    person_id = gets.chomp.to_i
-
-    print 'Date: '
-    date = gets.chomp.to_s
-
-    rental = Rental.new(date, @persons[person_id], @books[book_id])
-    @rentals << rental
-
-    puts 'Rental created successfully'
-  end
-
-  def list_all_rentals
-    print 'To see person rentals enter the person ID: '
-    id = gets.chomp.to_i
-
-    puts 'Rented Books:'
-    @rentals.each do |rental|
-      if rental.person.id == id
-        puts "Person: #{rental.person.name} Date: #{rental.date}, Book: '#{rental.book.title}' by #{rental.book.author}"
-
-      else
-        puts
-        puts 'No records where found for the given ID'
-      end
     end
   end
 end
